adding a simpler targetgraphbuilder
