+++ packages/target-graph/src/index.ts
@@ -5,4 +5,4 @@ export type { TargetConfig } from "./types/TargetConfig.js";
 export { sortTargetsByPriority } from "./sortTargetsByPriority.js";
 export { getPackageAndTask, getTargetId, getStartTargetId } from "./targetId.js";
 export { detectCycles } from "./detectCycles.js";
-export { TargetGraphBuilder } from "./TargetGraphBuilder.js";
+export { WorkspaceTargetGraphBuilder } from "./WorkspaceTargetGraphBuilder.js";
