+++ packages/target-graph/src/types/TargetGraph.ts
@@ -2,5 +2,4 @@ import type { Target } from "../types/Target.js";
 
 export interface TargetGraph {
   targets: Map<string, Target>;
-  dependencies: [string, string][];
 }
